Restructure code and add modules
This is first iteration of refactoring the code
