+++ app/prj.conf
@@ -32,3 +32,5 @@ CONFIG_PRINTK=y
 
 CONFIG_LOG=y
 # CONFIG_USB_DRIVER_LOG_LEVEL_ERR=y
+
+CONFIG_BT_RX_STACK_SIZE=4096
