+++ app/src/webusb.h
@@ -12,6 +12,9 @@
 #ifndef __WEBUSB_SERIAL_H__
 #define __WEBUSB_SERIAL_H__
 
+#include <zephyr/types.h>
+#include "command.h"
+
 /**
  * @brief Initializes WebUSB component
  *
@@ -32,6 +35,7 @@ int webusb_transmit(uint8_t *data, uint16_t size);
  *
  * @param [in] handlers Pointer to WebUSB command handler structure
  */
-void webusb_register_command_handler(void (*webusb_cmd_handler)(uint8_t *command_ptr, uint16_t command_length));
+void webusb_register_command_handler(void (*webusb_cmd_handler)(struct command_message *command_ptr,
+								uint16_t command_length));
 
 #endif /* __WEBUSB_SERIAL_H__ */
